isValidxxx methods location changed. SaveRecord running.
