+++ src/main/java/org/trainingTracker/servlets/SaveRecord.java
@@ -1,6 +1,7 @@
 package org.trainingTracker.servlets;
 
 import java.io.IOException;
+import java.util.List;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -9,10 +10,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import net.sf.json.JSONObject;
-import net.sf.json.JSONException;
+import net.sf.json.JSONArray;
 
 import org.trainingTracker.servlets.ServletCommon;
+import org.trainingTracker.database.dataAccesObject.ExercisesDAO;
 import org.trainingTracker.database.dataAccesObject.RecordsDAO;
+import org.trainingTracker.database.valueObject.ExerciseVO;
 import org.trainingTracker.database.valueObject.RecordVO;
 
 /**
@@ -69,9 +72,9 @@ public class SaveRecord extends HttpServlet {
         }
         
         // Field revision
-        if (!ServletCommon.isValidWeight(weight, response) |
-            !ServletCommon.isValidSeries(series, response) |
-            !ServletCommon.isValidRepetitions(repetitions, response)) {
+        if (!isValidWeight(weight, response) |
+            !isValidSeries(series, response) |
+            !isValidRepetitions(repetitions, response)) {
             response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
             error = true;
         }
@@ -81,8 +84,26 @@ public class SaveRecord extends HttpServlet {
                 // Creates an record in BD
                 if (RecordsDAO.addRecord(Integer.parseInt(exercise), user, Double.parseDouble(weight),
                                          Integer.parseInt(series), Integer.parseInt(repetitions))) {
+                    // Search for predifined exercises in BD
+                    JSONArray jsonExercises = new JSONArray();
+                    JSONObject jExercise, jRecord;
+                    List<RecordVO> list;
+                    
                     response.setStatus(HttpServletResponse.SC_OK);
-                    response.sendRedirect("listPerformed");
+                    for (ExerciseVO vo : ExercisesDAO.listUserExercises(user)) {
+                        jExercise = JSONObject.fromObject(vo.serialize());
+                        jExercise.remove("predefines");
+                        if(!(list=RecordsDAO.listRecords(user, vo.getId(), 1)).isEmpty()){
+                            jRecord = JSONObject.fromObject(list.get(0).serialize());
+                            jRecord.remove("exercise");
+                            jRecord.remove("nick");
+                            jRecord.remove("date");
+                            jExercise.putAll(jRecord);
+                        }
+                        jsonExercises.add(jExercise);
+                    }
+                    response.setContentType("application/json; charset=UTF-8");
+                    response.getWriter().write(jsonExercises.toString());
                 }
                 else {
                     response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
@@ -95,6 +116,87 @@ public class SaveRecord extends HttpServlet {
                 response.getWriter().println("Error interno en el servidor. Vuelva intentarlo más tarde");
             }
         }
+    }
+    
+    /**
+     * @param str
+     * @param response
+     * @returns true if str confirms weight specifications
+     */
+    static boolean isValidWeight (String str, HttpServletResponse response) throws IOException {
+        boolean error = false;
+        
+        try {
+            if (!(Double.parseDouble(str) > 0)) {
+                error = true;
+            }
+        }
+        catch (NullPointerException e) {
+            error = true;
+        }
+        catch (NumberFormatException e) {
+            error = true;
+        }
+        
+        if (error) {
+            response.getWriter().println("Peso no válido");
+        }
+        
+        return !error;
+    }
+    
+    /**
+     * @param str
+     * @param response
+     * @returns true if str confirms series specifications
+     */
+    static boolean isValidSeries (String str, HttpServletResponse response) throws IOException {
+        boolean error = false;
+        
+        try {
+            if (!(Integer.parseInt(str) > 0)) {
+                error = true;
+            }
+        }
+        catch (NullPointerException e) {
+            error = true;
+        }
+        catch (NumberFormatException e) {
+            error = true;
+        }
+        
+        if (error) {
+            response.getWriter().println("Número de series no válido");
+        }
+        
+        return !error;
+    }
+    
+    /**
+     * @param str
+     * @param response
+     * @returns true if str confirms repetitions specifications
+     */
+    static boolean isValidRepetitions (String str, HttpServletResponse response) throws IOException {
+        boolean error = false;
+        
+        try {
+            if (!(Integer.parseInt(str) > 0)) {
+                error = true;
+            }
+        }
+        catch (NullPointerException e) {
+            error = true;
+        }
+        catch (NumberFormatException e) {
+            error = true;
+        }
+        
+        if (error) {
+            response.getWriter().println("Número de repeticiones no válido");
+        }
+        
+        return !error;
     }
     
 }
