+++ src/main/java/org/trainingTracker/servlets/ServletCommon.java
@@ -27,86 +27,5 @@ class ServletCommon {
 		
 		return JSONObject.fromObject(jb.toString());
 	}
-	
-	/**
-	 * @param str
-	 * @param response
-	 * @returns true if str confirms weight specifications
-	 */
-	static boolean isValidWeight (String str, HttpServletResponse response) throws IOException {
-	    boolean error = false;
-	    
-	    try {
-	        if (!(Double.parseDouble(str) > 0)) {
-	            error = true;
-	        }
-	    }
-	    catch (NullPointerException e) {
-	        error = true;
-	    }
-	    catch (NumberFormatException e) {
-	        error = true;
-	    }
-	    
-	    if (error) {
-	        response.getWriter().println("Peso no válido");
-	    }
-	    
-	    return !error;
-	}
-    
-    /**
-     * @param str
-     * @param response
-     * @returns true if str confirms series specifications
-     */
-    static boolean isValidSeries (String str, HttpServletResponse response) throws IOException {
-        boolean error = false;
-        
-        try {
-            if (!(Integer.parseInt(str) > 0)) {
-                error = true;
-            }
-        }
-        catch (NullPointerException e) {
-            error = true;
-        }
-        catch (NumberFormatException e) {
-            error = true;
-        }
-        
-        if (error) {
-            response.getWriter().println("Número de series no válido");
-        }
-        
-        return !error;
-    }
-    
-    /**
-     * @param str
-     * @param response
-     * @returns true if str confirms repetitions specifications
-     */
-    static boolean isValidRepetitions (String str, HttpServletResponse response) throws IOException {
-        boolean error = false;
-        
-        try {
-            if (!(Integer.parseInt(str) > 0)) {
-                error = true;
-            }
-        }
-        catch (NullPointerException e) {
-            error = true;
-        }
-        catch (NumberFormatException e) {
-            error = true;
-        }
-        
-        if (error) {
-            response.getWriter().println("Número de repeticiones no válido");
-        }
-        
-        return !error;
-    }
 
 }
